refactored controlpanel
