+++ src/iosanita/policy/profiles/default/registry/settings.xml
@@ -2,7 +2,7 @@
 <registry xmlns:i18n="http://xml.zope.org/namespaces/i18n"
           i18n:domain="plone"
 >
-  <records interface="iosanita.policy.controlpanels.settings.IIoSanitaSettings" />
+  <records interface="iosanita.policy.interfaces.IIoSanitaSettings" />
 
   <!-- custom values -->
   <record field="allowed_sizes"
