+++ src/iosanita/policy/restapi/configure.zcml
@@ -3,10 +3,13 @@
     xmlns:plone="http://namespaces.plone.org/plone"
     >
 
-  <include
-      package="plone.restapi"
-      file="configure.zcml"
-      />
+  <include package="plone.restapi"/>
   <include package=".services" />
 
+  <!-- enable controlpanel on plone.restapi -->
+  <adapter
+      factory=".controlpanel.IoSanitaSettingsControlpanel"
+      provides="iosanita.policy.interfaces.IIoSanitaSettingsControlpanel"
+      name="iosanita-settings"
+      />
 </configure>
