+++ src/iosanita/policy/restapi/services/configure.zcml
@@ -6,10 +6,4 @@
 
   <include package=".search_tassonomie" />
 
-  <adapter
-      factory=".controlpanel.IoSanitaSettings"
-      provides="iosanita.policy.controlpanels.settings.IIoSanitaSettings"
-      name="iosanita-settings"
-      />
-
 </configure>
