+++ src/iosanita/policy/tests/test_controlpanel_api.py
@@ -26,6 +26,9 @@ class ControlpanelTest(unittest.TestCase):
     def test_controlpanel_exists(self):
         response = self.api_session.get("/@controlpanels/iosanita-settings")
 
+        import pdb
+
+        pdb.set_trace()
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.headers.get("Content-Type"), "application/json")
 
@@ -33,4 +36,4 @@ class ControlpanelTest(unittest.TestCase):
         response = self.api_session.get("/@controlpanels")
 
         titles = [x.get("title") for x in response.json()]
-        self.assertIn("IoSanita Settings", titles)
+        self.assertIn("Impostazioni IoSanita", titles)
